Fix storage errors for unit testing
In `InterpreterConfig.ContractValueHandler` function, we would override
the interpreter's shared state with a new storage based on emulator's
ledger.
This proved to be problematic, as the new storage would not contain
the composite values needed for testing (e.g. I.Test and I.Crypto).
To fix this, we simply store every slab from the new storage, to the
current environment's storage.